backend/requirements.txt
Django>=5.1.7,<5.2
djangorestframework>=3.15.2,<3.16
psycopg[binary]>=3.2.6,<3.3
drf-spectacular>=0.28.0,<0.29
Pillow>=11.1.0,<12
flake8>=7.1.2,<7.2
djangorestframework-simplejwt>=5.5.0,<5.6
django-extensions>=3.2.3,<3.3
django-debug-toolbar>=5.0.1,<5.1
django-cors-headers>=4.7.0,<4.8
channels>=4.2.0,<4.3
# djangochannelsrestframework>=1.3.0,<1.4
daphne>=4.1.2,<4.2
channels_redis>=4.2.1,<4.3
django-nested-admin>=4.1.1,<4.2
